[feat/#17] login, signup 액티비티, 데이터, 뷰모델 폴더 분리
