+++ app/src/main/java/com/sopt/now/presentation/auth/login/LoginState.kt
@@ -1,0 +1,7 @@
+package com.sopt.now.presentation.auth.login
+
+data class LoginState(
+    val isSuccess: Boolean,
+    val message: String,
+    val memberId: String? = null
+)
